% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlinkSeed-class.R
\name{dimnames,PlinkSeed-method}
\alias{dimnames,PlinkSeed-method}
\title{Method: dimnames for delayed plink}
\usage{
\S4method{dimnames}{PlinkSeed}(x)
}
\arguments{
\item{x}{PlinkSeed instance}
}
\value{
list of dimnames
}
\description{
Method: dimnames for delayed plink
}
