% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_example.R
\name{example_PlinkArray}
\alias{example_PlinkArray}
\title{produce PlinkArray from example data}
\usage{
example_PlinkArray(folder = tempdir())
}
\arguments{
\item{folder}{a path where unzipped example data will be managed}
}
\description{
produce PlinkArray from example data
}
\examples{
example_PlinkArray()
}
