% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlinkPyArraySeed-class.R
\name{extract_array,PlinkPyArraySeed-method}
\alias{extract_array,PlinkPyArraySeed-method}
\title{extractor}
\usage{
\S4method{extract_array}{PlinkPyArraySeed}(x, index)
}
\arguments{
\item{x}{seed instance}

\item{index}{list of suitable values for extracting elements}
}
\description{
extractor
}
