% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlinkArray-package.R
\docType{package}
\name{PlinkArray-package}
\alias{PlinkArray-package}
\title{PlinkArray: DelayedArray interface for plink bed files}
\description{
DelayedArray interface for plink bed files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vjcitn/DelayedPlink}
  \item Report bugs at \url{https://github.com/vjcitn/DelayedPlink/issues}
}

}
\author{
\strong{Maintainer}: Vince Carey \email{stvjc@channing.harvard.edu} (\href{https://orcid.org/0000-0003-4046-0063}{ORCID})

}
\keyword{internal}
