% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlinkSeed-class.R
\name{PlinkArray}
\alias{PlinkArray}
\title{Constructor for DelayedArray}
\usage{
PlinkArray(filepath)
}
\arguments{
\item{filepath}{path to plink bed, bim, fam resources without suffixes}
}
\description{
Constructor for DelayedArray
}
