% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlinkMatrix-package.R
\docType{package}
\name{PlinkMatrix-package}
\alias{PlinkMatrix-package}
\title{PlinkMatrix: DelayedArray interface for plink bed files}
\description{
This package provides a DelayedArray interface for plink bed files. There is support for interfacing to plink genotype data via RangedSummarizedExperiment. Example data from the GEUVADIS project (internationalgenome.org) are used for demonstration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vjcitn/PlinkMatrix}
  \item Report bugs at \url{https://github.com/vjcitn/PlinkMatrix/issues}
}

}
\author{
\strong{Maintainer}: Vince Carey \email{stvjc@channing.harvard.edu} (\href{https://orcid.org/0000-0003-4046-0063}{ORCID})

Other contributors:
\itemize{
  \item NHGRI U24HG004059 [funder]
}

}
\keyword{internal}
