% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlinkPyArraySeed-class.R
\name{PlinkPyArray}
\alias{PlinkPyArray}
\title{bed reader interface}
\usage{
PlinkPyArray(path)
}
\arguments{
\item{path}{character path to resource}
}
\description{
bed reader interface
}
\examples{
folder = tempdir()
pa = get_plink_example_path()
unzip(pa, exdir = folder)
tmp = dir(folder, full=TRUE)
ppath = grep("geuv445.bed", tmp, value=TRUE)
tst = PlinkPyArray(ppath)
tst
}
