% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlinkPyArraySeed-class.R
\name{PlinkPyArraySeed}
\alias{PlinkPyArraySeed}
\title{seed}
\usage{
PlinkPyArraySeed(filepath, readerRef)
}
\arguments{
\item{filepath}{character path to resource}

\item{readerRef}{reticulate-imported `bed_reader` module}
}
\description{
seed
}
