% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlinkSeed-class.R
\docType{class}
\name{PlinkSeed-class}
\alias{PlinkSeed-class}
\title{Define the PlinkSeed class}
\value{
class definition
}
\description{
Define the PlinkSeed class
}
\examples{
getClass("PlinkSeed")
}
\keyword{internal}
