% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlinkSeed-class.R
\name{PlinkSeed}
\alias{PlinkSeed}
\title{Constructor function for seed for plink bed format}
\usage{
PlinkSeed(filepath)
}
\arguments{
\item{filepath}{character string without suffixes}
}
\value{
PlinkSeed instance
}
\description{
Constructor function for seed for plink bed format
}
\examples{
PlinkSeed
}
\keyword{internal}
