% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlinkPyArraySeed-class.R
\name{show,PlinkPyArraySeed-method}
\alias{show,PlinkPyArraySeed-method}
\title{present seed concisely}
\usage{
\S4method{show}{PlinkPyArraySeed}(object)
}
\arguments{
\item{object}{instance of PlinkPyArraySeed}
}
\description{
present seed concisely
}
