% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangedPlink.R
\name{show,RangedPlink-method}
\alias{show,RangedPlink-method}
\title{concise presentation}
\usage{
\S4method{show}{RangedPlink}(object)
}
\arguments{
\item{object}{RangedPlink instance}
}
\description{
concise presentation
}
