% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangedPlink.R
\name{subsetByOverlaps,RangedPlink,GRanges-method}
\alias{subsetByOverlaps,RangedPlink,GRanges-method}
\title{implement a form of subsetting}
\usage{
\S4method{subsetByOverlaps}{RangedPlink,GRanges}(
  x,
  ranges,
  maxgap = -1L,
  minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"),
  invert = FALSE,
  ...
)
}
\arguments{
\item{x}{instance of RangedPlink}

\item{ranges}{instance of GenomicRanges}

\item{maxgap}{integer, see subsetByOverlaps, defaults to -1L}

\item{minoverlap}{integer, see subsetByOverlaps, defaults to 0L}

\item{type}{character, see subsetByOverlaps, defaults to 'any'}

\item{invert, }{logical, see subsetByOverlaps}

\item{\dots}{not used}
}
\value{
RangedPlink instance
}
\description{
implement a form of subsetting
}
