% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_bed_genotypes.R
\name{decode_bed_genotypes}
\alias{decode_bed_genotypes}
\title{Helper function to decode BED format}
\usage{
decode_bed_genotypes(raw_bytes, n_samples)
}
\arguments{
\item{raw_bytes}{byte stream from file}

\item{n_samples}{number of samples}
}
\description{
Helper function to decode BED format
}
\keyword{internal}
