% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlinkSeed-class.R
\name{dim,PlinkSeed-method}
\alias{dim,PlinkSeed-method}
\title{Method: dim for delayed plink}
\usage{
\S4method{dim}{PlinkSeed}(x)
}
\arguments{
\item{x}{PlinkSeed instance}
}
\value{
2-vector
}
\description{
Method: dim for delayed plink
}
\examples{
methods(class="PlinkSeed")
}
