% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_example.R
\name{example_PlinkMatrix}
\alias{example_PlinkMatrix}
\title{produce PlinkMatrix from example data}
\usage{
example_PlinkMatrix(folder = tempdir())
}
\arguments{
\item{folder}{a path where unzipped example data will be managed}
}
\description{
produce PlinkMatrix from example data
}
\examples{
example_PlinkMatrix()
}
