% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayedPlink.R
\name{extract_array,PlinkSeed-method}
\alias{extract_array,PlinkSeed-method}
\title{Method: extract_array, internal}
\usage{
\S4method{extract_array}{PlinkSeed}(x, index)
}
\description{
Method: extract_array, internal
}
