% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g445samples}
\alias{g445samples}
\title{sample characteristics of 445 GEUVADIS samples}
\format{
data.frame
}
\usage{
data("g445samples", package="PlinkMatrix")
}
\description{
sample characteristics of 445 GEUVADIS samples
}
\note{
Example data are those provided with tensorqtl, see \url{https://github.com/broadinstitute/tensorqtl/tree/0c4db65a0cdc47f3b824ae530b89d270ef5e0096/example/data}.
}
\keyword{datasets}
