% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g445samples}
\alias{g445samples}
\title{sample characteristics of 445 GEUVADIS samples}
\format{
data.frame
}
\usage{
data("g445samples", package="PlinkArray")
}
\description{
sample characteristics of 445 GEUVADIS samples
}
\keyword{datasets}
