% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_example.R
\name{get_plink_example_path}
\alias{get_plink_example_path}
\title{operate with BiocFileCache to retrieve zip file of plink example data}
\usage{
get_plink_example_path(ca = BiocFileCache::BiocFileCache())
}
\arguments{
\item{ca}{BiocFileCache instance}
}
\description{
operate with BiocFileCache to retrieve zip file of plink example data
}
\examples{
get_plink_example_path()
}
