% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plinkHelp.R
\name{getSampleData}
\alias{getSampleData}
\title{Get sample metadata}
\usage{
getSampleData(x)
}
\arguments{
\item{x}{DelayedArray instance}
}
\description{
Get sample metadata
}
\examples{
tst = example_PlinkArray()
head(getSampleData(tst))
}
