% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plinkHelp.R
\name{getVariantData}
\alias{getVariantData}
\title{Get variant metadata}
\usage{
getVariantData(x)
}
\arguments{
\item{x}{DelayedArray instance
tst = example_PlinkArray()
dim(getVariantData(tst))}
}
\description{
Get variant metadata
}
