% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plinkHelp.R
\name{getVariantData}
\alias{getVariantData}
\title{Get variant metadata}
\usage{
getVariantData(x)
}
\arguments{
\item{x}{DelayedArray instance}
}
\description{
Get variant metadata
}
\examples{
tst = example_PlinkMatrix()
dim(getVariantData(tst))
}
