% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plid2gr.R
\name{plid2gr}
\alias{plid2gr}
\title{produce GRanges from variant notation for plink example from geuvadis}
\usage{
plid2gr(x, sepused = "_")
}
\arguments{
\item{x}{character vector of variant names}

\item{sepused}{single character, defaults to "_"}
}
\value{
GRanges instance
}
\description{
produce GRanges from variant notation for plink example from geuvadis
}
