% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlinkMatrix-package.R
\name{PlinkMatrix-pkg}
\alias{PlinkMatrix-pkg}
\title{Delayed interface to Plink genotype files}
\description{
This package includes C++ code to interface to Plink files.
Large-scale genotype calls can be managed in a RangedSummarized
Experiment instance.
}
