% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_bed_subset}
\alias{read_bed_subset}
\title{read subset for use in delayed matrix approach}
\usage{
read_bed_subset(prefix, snp_indices, sample_indices, n_total_samples = NULL)
}
\arguments{
\item{prefix}{character path to bed resources and file prefix}

\item{snp_indices}{integer vector of snp indices}

\item{sample_indices}{integer vector of sample indices}

\item{n_total_samples}{optional}
}
\description{
read subset for use in delayed matrix approach
}
\examples{
read_bed_subset
}
