% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlinkSeed-class.R
\name{show,PlinkSeed-method}
\alias{show,PlinkSeed-method}
\title{present seed concisely}
\usage{
\S4method{show}{PlinkSeed}(object)
}
\arguments{
\item{object}{instance of PlinkSeed}
}
\value{
side effect of cat
}
\description{
present seed concisely
}
\examples{
ex <- example_PlinkMatrix()
ex
}
